+++ news-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## DEV
 
+ * Link to the news article if no image link has been defined (see #30).
  * Show the 404 page if an external news is opened in the news reader (see #33).